Cache static build too
