+++ src/builder/webpack/config.common.js
@@ -3,18 +3,27 @@
 import { DefinePlugin } from "webpack"
 import url from "url"
 import pkg from "../../../package.json"
+import {
+  getCacheDir,
+  hardSourceRecordsPath,
+  hardSourcePlugin,
+} from "../../_utils/hard-source-webpack-plugin"
 
 export const chunkNameBrowser = "phenomic.browser"
 
 const wrap = JSON.stringify
 
 export default (config: PhenomicConfig): WebpackConfig => {
+
+  const cacheDir = getCacheDir(config)
   const { webpackConfig = {} } = config
 
   return {
     ...webpackConfig,
+    ...config.webpackHardCache && hardSourceRecordsPath(cacheDir),
     plugins: [
       ...webpackConfig.plugins,
+      ...config.webpackHardCache && hardSourcePlugin(cacheDir, config),
 
       new DefinePlugin({ "process.env": {
         NODE_ENV: wrap(
